수정: 클래스 계층, 이름 수정
