+++ app/services/external_channel/order/saver.rb
@@ -1,7 +1,7 @@
 module ExternalChannel
   module Order
     # TODO: 방어 로직 추가
-    class Saver < DataSaver
+    class Saver < BaseSaver
       def save_all(orders)
         orders.all? {|order| save(order)}
       end
