+++ app/services/external_channel/product/manager.rb
@@ -1,6 +1,6 @@
 module ExternalChannel
   module Product
-    class Service < DataService
+    class Manager < BaseManager
       def initialize(channel_adapter)
         super(channel_adapter)
         set_saver!(ExternalChannel::Product::Saver.new)
